requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
pymysql==1.1.0
cryptography==42.0.5
python-dotenv==1.0.1
Werkzeug==3.0.1
